refactor: move a function and improve logging in prep for persisting cache
